Rename TestUtils to UnitTestRegistration. (#4021)
Looking to remove usage of 'Utils' unless we have really no choice.
'UnitTestRegistration' seems clearer in what it does compared to
'TestUtils'.
