+++ src/ble/tests/TestBleErrorStr.cpp
@@ -40,7 +40,7 @@
 
 #include <ble/BleError.h>
 #include <support/ErrorStr.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
