+++ src/ble/tests/TestBleUUID.cpp
@@ -27,7 +27,7 @@
 #include "TestBleLayer.h"
 
 #include <ble/BleUUID.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
