+++ src/inet/tests/TestInetAddress.cpp
@@ -46,7 +46,7 @@
 #include <inet/IPPrefix.h>
 
 #include <support/CodeUtils.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 using namespace chip;
 using namespace chip::Inet;
