+++ src/inet/tests/TestInetEndPoint.cpp
@@ -40,7 +40,7 @@
 
 #include <support/CHIPArgParser.hpp>
 #include <support/CodeUtils.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <system/SystemError.h>
 #include <system/SystemTimer.h>
