+++ src/inet/tests/TestInetErrorStr.cpp
@@ -39,7 +39,7 @@
 #include <inet/InetError.h>
 #include <support/CodeUtils.h>
 #include <support/ErrorStr.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
