+++ src/inet/tests/TestInetLayerDNS.cpp
@@ -38,7 +38,7 @@
 
 #include <inet/InetLayer.h>
 #include <support/CodeUtils.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <system/SystemClock.h>
 #include <system/SystemTimer.h>
