+++ src/lib/core/tests/TestCHIPCallback.cpp
@@ -26,7 +26,7 @@
 
 #include <core/CHIPCallback.h>
 #include <support/CHIPMem.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
