+++ src/lib/core/tests/TestCHIPErrorStr.cpp
@@ -40,7 +40,7 @@
 
 #include <core/CHIPError.h>
 #include <support/ErrorStr.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
