+++ src/lib/core/tests/TestCHIPTLV.cpp
@@ -39,7 +39,7 @@
 #include <support/CodeUtils.h>
 #include <support/RandUtils.h>
 #include <support/ScopedBuffer.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <string.h>
