+++ src/lib/core/tests/TestReferenceCounted.cpp
@@ -30,7 +30,7 @@
 #include <string.h>
 
 #include <lib/core/ReferenceCounted.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
