+++ src/lib/mdns/minimal/tests/TestIPResourceRecord.cpp
@@ -17,7 +17,7 @@
  */
 
 #include <mdns/minimal/IPResourceRecord.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
