+++ src/lib/mdns/minimal/tests/TestQName.cpp
@@ -17,7 +17,7 @@
  */
 
 #include <mdns/minimal/QName.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
