+++ src/lib/mdns/minimal/tests/TestRecordData.cpp
@@ -19,7 +19,7 @@
 #include <string>
 #include <vector>
 
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
