+++ src/lib/mdns/minimal/tests/TestResourceRecord.cpp
@@ -17,7 +17,7 @@
  */
 
 #include <mdns/minimal/ResourceRecord.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
