+++ src/lib/support/BUILD.gn
@@ -73,10 +73,10 @@ static_library("support") {
     "SafeInt.h",
     "SerializableIntegerSet.cpp",
     "SerializableIntegerSet.h",
-    "TestUtils.cpp",
-    "TestUtils.h",
     "TimeUtils.cpp",
     "TimeUtils.h",
+    "UnitTestRegistration.cpp",
+    "UnitTestRegistration.h",
     "logging/CHIPLogging.cpp",
     "logging/CHIPLogging.h",
     "logging/CHIPLoggingLogV.cpp",
