+++ src/lib/support/tests/TestBufBound.cpp
@@ -25,7 +25,7 @@
 #include "TestSupport.h"
 
 #include <support/BufBound.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
