+++ src/lib/support/tests/TestBufferReader.cpp
@@ -25,7 +25,7 @@
 #include "TestSupport.h"
 
 #include <support/BufferReader.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 #include <type_traits>
 
 #include <nlunit-test.h>
