+++ src/lib/support/tests/TestSafeInt.cpp
@@ -25,7 +25,7 @@
 #include "TestSupport.h"
 
 #include <support/SafeInt.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
