+++ src/lib/support/tests/TestSerializableIntegerSet.cpp
@@ -21,7 +21,7 @@
 #include <support/CHIPMem.h>
 #include <support/CHIPMemString.h>
 #include <support/SerializableIntegerSet.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <nlunit-test.h>
