+++ src/platform/tests/TestConfigurationMgr.cpp
@@ -33,7 +33,7 @@
 #include <nlunit-test.h>
 #include <support/CHIPMem.h>
 #include <support/CodeUtils.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <platform/CHIPDeviceLayer.h>
