+++ src/platform/tests/TestPlatformTime.cpp
@@ -32,7 +32,7 @@
 
 #include <nlunit-test.h>
 #include <support/CodeUtils.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 #include <system/SystemClock.h>
 
 #include <platform/internal/CHIPDeviceLayerInternal.h>
