+++ src/protocols/bdx/tests/TestBdxMessages.cpp
@@ -4,7 +4,7 @@
 
 #include <support/BufBound.h>
 #include <support/CodeUtils.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <limits>
