+++ src/system/tests/TestSystemPacketBuffer.cpp
@@ -35,7 +35,7 @@
 #include <string.h>
 
 #include <support/CodeUtils.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 #include <system/SystemPacketBuffer.h>
 
 #if CHIP_SYSTEM_CONFIG_USE_LWIP
