+++ src/system/tests/TestSystemTimer.cpp
@@ -35,7 +35,7 @@
 #include <nlunit-test.h>
 #include <support/CodeUtils.h>
 #include <support/ErrorStr.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 #include <system/SystemError.h>
 #include <system/SystemLayer.h>
 #include <system/SystemTimer.h>
