+++ src/system/tests/TestTimeSource.cpp
@@ -33,7 +33,7 @@
 #include <nlunit-test.h>
 #include <support/CodeUtils.h>
 #include <support/ErrorStr.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 #include <system/TimeSource.h>
 
 namespace {
