+++ src/test_driver/esp32/main/main_app.cpp
@@ -31,7 +31,7 @@
 #include <crypto/CHIPCryptoPAL.h>
 #include <platform/CHIPDeviceLayer.h>
 #include <support/ErrorStr.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 using namespace ::chip;
 using namespace ::chip::DeviceLayer;
