+++ src/test_driver/nrfconnect/main/runner.cpp
@@ -18,7 +18,7 @@
 #include <platform/CHIPDeviceLayer.h>
 #include <support/CodeUtils.h>
 #include <support/ErrorStr.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 
 #include <logging/log.h>
 #include <settings/settings.h>
