+++ src/transport/raw/tests/TestMessageHeader.cpp
@@ -24,7 +24,7 @@
  */
 #include <support/CodeUtils.h>
 #include <support/ErrorStr.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 #include <transport/raw/MessageHeader.h>
 
 #include <nlunit-test.h>
