+++ src/transport/tests/TestPeerConnections.cpp
@@ -26,7 +26,7 @@
 
 #include <support/CodeUtils.h>
 #include <support/ErrorStr.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 #include <transport/PeerConnections.h>
 
 #include <nlunit-test.h>
