+++ src/transport/tests/TestSecurePairingSession.cpp
@@ -31,7 +31,7 @@
 #include <stdarg.h>
 #include <support/CHIPMem.h>
 #include <support/CodeUtils.h>
-#include <support/TestUtils.h>
+#include <support/UnitTestRegistration.h>
 #include <transport/SecurePairingSession.h>
 
 using namespace chip;
